It doesn't work yes, Need to deserialize from double to long

+++ frontend_android/app/src/main/java/pl/pollub/harnasik/app/domain/model/OfferSpot.kt
@@ -1,7 +1,7 @@
 package pl.pollub.harnasik.app.domain.model
 
 data class OfferSpot(
-    val lat: Double,
-    val lng: Double,
+    val lat: Long,
+    val lng: Long,
     val id: Int? = null
 )
